refactor tests into own files
