+++ tests/functional/adapter/basic/test_adapter_methods.py
@@ -1,0 +1,86 @@
+import pytest
+from dbt.tests.adapter.basic.test_adapter_methods import BaseAdapterMethod
+from dbt.tests.adapter.basic.test_adapter_methods import models__upstream_sql
+
+models__my_model_sql = """
+
+{% set upstream = ref('upstream_view') %}
+
+{% if execute %}
+    {# don't ever do any of this #}
+    {%- do adapter.drop_schema(upstream) -%}
+    {% set existing = adapter.get_relation(upstream.database, upstream.schema, upstream.identifier) %}
+    {% if existing is not defined %}
+        {% do exceptions.raise_compiler_error('expected ' ~ ' to not exist, but it did') %}
+    {% endif %}
+
+    {%- do adapter.create_schema(upstream) -%}
+
+    {% set sql = create_view_as(upstream, 'select 2 as id') %}
+    {% do run_query(sql) %}
+{% endif %}
+
+
+select * from {{ upstream }}
+
+"""
+
+models__expected_sql = """
+-- make sure this runs after 'model'
+-- {{ ref('model_view') }}
+select 2 as id
+
+"""
+
+
+class TestBaseAdapterMethodDremio(BaseAdapterMethod):
+    @pytest.fixture(scope="class")
+    def project_config_update(self):
+        return {
+            "models": {
+                "+twin_strategy": "clone",
+            },
+            "name": "adapter_methods",
+        }
+
+    @pytest.fixture(scope="class")
+    def models(self):
+        return {
+            "upstream_view.sql": models__upstream_sql,
+            "expected_view.sql": models__expected_sql,
+            "model_view.sql": models__my_model_sql,
+        }
+
+    @pytest.fixture(scope="class")
+    def unique_schema(self, request, prefix) -> str:
+        test_file = request.module.__name__
+        # We only want the last part of the name
+        test_file = test_file.split(".")[-1]
+        unique_schema = f"rav-test.{prefix}_{test_file}"
+        return unique_schema
+
+    @pytest.fixture(scope="class")
+    def dbt_profile_data(
+        self, unique_schema, dbt_profile_target, profiles_config_update
+    ):
+        profile = {
+            "config": {"send_anonymous_usage_stats": False},
+            "test": {
+                "outputs": {
+                    "default": {},
+                },
+                "target": "default",
+            },
+        }
+        target = dbt_profile_target
+        target["schema"] = unique_schema
+        target["root_path"] = unique_schema
+        profile["test"]["outputs"]["default"] = target
+
+        if profiles_config_update:
+            profile.update(profiles_config_update)
+        return profile
+
+    @pytest.fixture(scope="class")
+    def equal_tables(self):
+        return ["model_view", "expected_view"]
